requirements.txt
customtkinter
ffmpeg-python
google-generativeai
Pillow
google-genai
